yay is enough to install things
using only yay for installations.

divide packages into to part: explicitly installed and optional for explicit packages. so if you want to install a package you should add its name to the explicit packages and add its optional dependencies into optional package list. I thinks it is better because the automatic installation of optional packages did not work good and also installation of all the optional packages may be not a good idea.
enjoy it.